Check linter errors
